requirements.txt
Flask==2.2.5
scikit-learn==1.6.1
pandas==1.5.3
numpy==1.26.4
joblib==1.4.2
matplotlib==3.9.4
seaborn==0.13.2
dvc==3.59.0
PyYAML==6.0.2
web3==6.10.0
py-solc-x==1.1.0
eth-account==0.10.0
python-dotenv==1.0.0
